pyproject.toml
[project]
name = "skillsync"
version = "0.1.0"
description = "RAG-based curriculum gap analyzer for CS/IT departments (NBA SAR 2025)"
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
    # --- Core ---
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "loguru>=0.7.2",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",               # Retry logic for API rate limits

    # --- Document processing ---
    "pymupdf4llm>=0.0.17",
    "pymupdf>=1.24.0",
    "pandas>=2.2.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",

    # --- NLP / ML (CPU-friendly versions) ---
    "spacy>=3.7.4",
    "sentence-transformers>=3.0.0",
    "transformers>=4.41.0",
    "torch>=2.3.0",                  # CPU wheel — Windows installs it automatically
    "scikit-learn>=1.5.0",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.9.0",
    "numpy>=1.26.0",

    # --- LLM clients (cloud-first) ---
    "groq>=0.9.0",                   # Official Groq client
    "google-generativeai>=0.7.0",    # Gemini client
    "ollama>=0.2.1",                 # Local fallback

    # --- LangChain / LangGraph ---
    "langchain>=0.2.5",
    "langchain-community>=0.2.5",
    "langchain-core>=0.2.9",
    "langchain-groq>=0.1.5",
    "langchain-google-genai>=1.0.6",
    "langchain-ollama>=0.1.0",
    "langchain-qdrant>=0.1.1",
    "langgraph>=0.1.5",

    # --- Vector / Storage / Cache ---
    "qdrant-client>=1.9.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "redis>=5.0.4",
    "minio>=7.2.7",

    # --- Observability ---
    "langfuse>=2.36.0",

    # --- Structured outputs ---
    "instructor>=1.3.3",

    # --- API / UI ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.35.0",
    "reportlab>=4.2.0",
    "jinja2>=3.1.4",

    # --- Evaluation ---
    "ragas>=0.1.9",
    "datasets>=2.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.8",
    "mypy>=1.10.0",
    "ipykernel>=6.29.0",
    "jupyterlab>=4.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
